requirements.txt
MLB-StatsAPI==1.4.2
plotly==5.8.0
matplotlib-venn==0.11.7
python-mlb-statsapi==0.5.12